requirements.txt
requests
sqlmodel
fastapi
xmltodict
psycopg2-binary
python-dotenv
loguru
notifiers==1.3.3
ShopifyAPI
graphene
